refactor(types): move recurrence util into shared package, add occursOn
