+++ packages/types/package.json
@@ -6,5 +6,8 @@
     ".": "./src/index.ts"
   },
   "private": true,
-  "types": "src/index.ts"
+  "types": "src/index.ts",
+  "dependencies": {
+    "date-fns": "^4.1.0"
+  }
 }
