+++ packages/types/src/index.ts
@@ -111,3 +111,5 @@ export interface PushSubscription {
     };
     dateAdded: Date;
 }
+
+export { expandOccurrences, isoDay, type Occurrence, occursOn } from './recurrence';
